Moving to 8.1.0

+++ RedProtect-Spigot/src/main/java/br/net/fabiozumbi12/RedProtect/Bukkit/actions/DefineRegionBuilder.java
@@ -177,7 +177,7 @@ public class DefineRegionBuilder extends RegionBuilder {
         long reco = 0;
         if (RedProtect.get().getConfigManager().ecoRoot().claim_cost_per_block.enable && RedProtect.get().hooks.checkVault() && !p.hasPermission("redprotect.eco.bypass")) {
             double peco = RedProtect.get().economy.getBalance(p);
-            reco = newRegion.getArea() * RedProtect.get().getConfigManager().ecoRoot().claim_cost_per_block.cost_per_block;
+            reco = (long) newRegion.getArea() * RedProtect.get().getConfigManager().ecoRoot().claim_cost_per_block.cost_per_block;
 
             if (!RedProtect.get().getConfigManager().ecoRoot().claim_cost_per_block.y_is_free) {
                 reco = reco * Math.abs(newRegion.getMaxY() - newRegion.getMinY());
